third_party/blink/web_tests/fast/dom/shadow/iframe-shadow-expected.txt

This test ensures that iframe inside shadow isn't visible from the host document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.length is 4
PASS i0byname.length is 1
PASS i0 is i0byname[0].contentWindow
PASS i1.length is 2
PASS window.length is 4
PASS i2byname.length is 1
PASS i2byname[0].contentWindow is i2
PASS i3byname.length is 0
PASS i3 is undefined
PASS window.length is 3
PASS document['i0'] is undefined
PASS window.length is 4
PASS document['i3'] is i3InShadow.contentWindow
PASS successfullyParsed is true

TEST COMPLETE

  
